Renaming and adding docstrings to public module functions
